Fixing the frontend to properly display the build information (#22)
Some fixes to the UI so that it properly displays the build information. 
Previously, it wasn't dropping down when the 'About' was clicked, and the Sign-in and Sign-up buttons were squished and not right-aligned.
Also, didn't test it but the buildinfo page when navigated to manually likely wasn't loading properly either, so that's been addressed as well.